google_cloud_function_download_youtube_video/requirements.txt
# Function dependencies, for example:
# package>=version
pytube
google-cloud-storage
moviepy
